# perf(vector): updated marshalling of vector