--- a/dgraphtest/config.go
+++ b/dgraphtest/config.go
@@ -147,14 +147,14 @@ func NewClusterConfig() ClusterConfig {
 	}
 }
 
-func newClusterConfigFrom(cc ClusterConfig) ClusterConfig {
-	prefix := fmt.Sprintf("dgraphtest-%d", rand.NewSource(time.Now().UnixNano()).Int63()%1000000)
-	defaultBackupVol := fmt.Sprintf("%v_backup", prefix)
-	defaultExportVol := fmt.Sprintf("%v_export", prefix)
-	cc.prefix = prefix
-	cc.volumes = map[string]string{DefaultBackupDir: defaultBackupVol, DefaultExportDir: defaultExportVol}
-	return cc
-}
+//func newClusterConfigFrom(cc ClusterConfig) ClusterConfig {
+//	prefix := fmt.Sprintf("dgraphtest-%d", rand.NewSource(time.Now().UnixNano()).Int63()%1000000)
+//	defaultBackupVol := fmt.Sprintf("%v_backup", prefix)
+//	defaultExportVol := fmt.Sprintf("%v_export", prefix)
+//	cc.prefix = prefix
+//	cc.volumes = map[string]string{DefaultBackupDir: defaultBackupVol, DefaultExportDir: defaultExportVol}
+//	return cc
+//}
 
 // WithNAlphas sets the number of alphas in the cluster
 func (cc ClusterConfig) WithNumAlphas(n int) ClusterConfig {