--- a/posting/mvcc.go
+++ b/posting/mvcc.go
@@ -535,7 +535,7 @@ func (c *CachePL) Set(l *List, readTs uint64) {
 }
 
 func ShouldGoInCache(pk x.ParsedKey) bool {
-	return !pk.IsData() && strings.HasSuffix(pk.Attr, "dgraph.type")
+	return (!pk.IsData() && strings.HasSuffix(pk.Attr, "dgraph.type"))
 }
 
 func getNew(key []byte, pstore *badger.DB, readTs uint64) (*List, error) {